Created a ImageFade script and attached it to the DayCycle Component
The ImageFade script allows for any UI image, to fade from 0->1 or fade from 1->0. Additionally, it allows for callbacks that will get called when a transiation is complete.
Finally, I added it to the day cycle, so that when the level starts it fades in, when the level ends it fades out, and when there's a day segment change, it fades out and then fades in