Modifying the CinemaCam and refacatored the Scriptable Animations
After using the CinemaCam for a bit, I realised how impractical it is, and therefor I'm slightly modifying it to be more modular, making the animation process quicker.

Additionally, I made a small adjustment to the scriptable animations as I realised they were being used for doors and other general animations. Now it is its own dedicated ScriptableAnimation, making it 100% seperate from the CinemaDetails (however, it is still a very crucial part of it)