Added almost all SFX to day 1, updated the minigame framework, & adjustm
Added SFX to practically everything in day 1.
Including added the ability for games to play audio-clips on the TV, making sure that all sounds from the mini games come out of the TV using 3D sound.
As well as added the ability to have a theme song in the game - Turning off the TV does not stop the song, but rather just sets its volume to 0; i.e.  the game doesn't stop because the tv is off :smiley:
by SFX, I mean I added them to the doors/drawers, as well as objects colliding with the ground as well as being picked up.

Additionally, I greated a general purpose Toggler as opposed to the EventToggler, since it becamse very tedious to add audio animations and events to the appropriate things